test: WinningResult 테스트 추가
